Updated the getPlaylist and created an interactor for it.
